fix: CI green — clippy, WASM clippy, gofmt, release date
- `#[derive(Default)]` on `RenderingIntent` with `#[default]` marker;
  drops the manual impl clippy flagged as derivable.
- Drop redundant `as usize` cast on `char_count` in xycut (already
  usize from `.count().max(1)`).
- `std::iter::repeat_n(v, n)` in place of `std::iter::repeat(v).take(n)`
  in predictor tests and the 4bpc synthetic fixture (clippy
  `manual_repeat_n` flagged as stabilised since 1.82).
- `gofmt -w go/pdf_oxide.go` to appease the Go format-check job.
- Bump v0.3.34 release date in CHANGELOG to 2026-04-18 (merge day).

Verified locally:
- `cargo clippy --all-targets --workspace -- -D warnings` — clean
- `cargo clippy --target wasm32-unknown-unknown --no-default-features \
    --features wasm --lib -- -D warnings` — clean
- `cargo fmt --check` — clean
- `gofmt -l go/` — clean

+++ src/pipeline/reading_order/xycut.rs
@@ -548,7 +548,7 @@ impl XYCutStrategy {
             // like 'G', 'T', '1', 'A') that scatter across the full X range
             // and fill the column gutter in the density profile. Body text
             // spans always contain multiple characters.
-            if (char_count as usize) < 2 {
+            if char_count < 2 {
                 continue;
             }
             let core_left = span.bbox.left();
